+++ tt_metal/api/tt-metalium/experimental/sockets/mesh_socket.hpp
@@ -161,20 +161,7 @@ public:
 
     tt::tt_fabric::FabricNodeId get_fabric_node_id(SocketEndpoint endpoint, const MeshCoordinate& coord) const;
 
-    std::vector<MeshCoreCoord> get_active_cores() const {
-        std::vector<MeshCoreCoord> active_cores;
-        active_cores.reserve(config_.socket_connection_config.size());
-        if (socket_endpoint_type_ == SocketEndpoint::SENDER) {
-            for (const auto& connection : config_.socket_connection_config) {
-                active_cores.push_back(connection.sender_core);
-            }
-        } else {
-            for (const auto& connection : config_.socket_connection_config) {
-                active_cores.push_back(connection.receiver_core);
-            }
-        }
-        return active_cores;
-    };
+    std::vector<MeshCoreCoord> get_active_cores() const;
 
     MeshDevice* get_mesh_device() const { return config_buffer_->device(); }
     static constexpr auto attribute_names =
