+++ tt_metal/distributed/mesh_socket.cpp
@@ -276,6 +276,21 @@ tt::tt_fabric::FabricNodeId MeshSocket::get_fabric_node_id(SocketEndpoint endpoi
     return fabric_node_id_map_[static_cast<std::underlying_type_t<SocketEndpoint>>(endpoint)].at(coord);
 }
 
+std::vector<MeshCoreCoord> MeshSocket::get_active_cores() const {
+    std::vector<MeshCoreCoord> active_cores;
+    active_cores.reserve(config_.socket_connection_config.size());
+    if (socket_endpoint_type_ == SocketEndpoint::SENDER) {
+        for (const auto& connection : config_.socket_connection_config) {
+            active_cores.push_back(connection.sender_core);
+        }
+    } else {
+        for (const auto& connection : config_.socket_connection_config) {
+            active_cores.push_back(connection.receiver_core);
+        }
+    }
+    return active_cores;
+}
+
 }  // namespace tt::tt_metal::distributed
 
 namespace std {
